separate dispatcher middleware from dispatcher
